Spikes do damage, unstable platforms shrink.
The spikes now do damage and the unstable platforms now shrink before dissapearing. The retracable spikes are a bit buggy but "work." Sometimes if the player drops down directly on them the spike will do damage and bounce the player before being fully retracted and sometimes nothing happens. I'll keep working on debugging it but for now the retractable spikes aren't part of the prefab.